+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Reworked the built-in theme
 - [moth.mjs](theme/moth.mjs) is now the standard MOTH library for ECMAScript
 - Devel mode no longer accepts an empty team ID
+- messages.html moved into theme
+- Exported state now includes "Enabled" boolean
 
 ## [v4.4.9] - 2022-05-12
 ### Changed